+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 
 ## vNext
+* DNS: Adds private resolver and forwarding rulesets
 * PostgreSQL: Adds `FullyQualifiedDomainName` configuration member.
 
 ## 1.7.18